use accept_data instead combine_validity for setting position_valid
In case of surface global CPR one CPR can be up to 50 seconds old
leading to unlogically short expiration of the resulting position.
Instead use accept_data to update position_valid when a new position is
dedoced.
Also check the result of accept_data to decide if we actually want to
update the position.
This is important if we do local CPR from a synthetic MLAT message.
One ADS-B CPR and the positon might not be stale when the position is
from a local CPR decode. Now we avoid overriding the ADS-B position if
it's not stale yet, this is exactly what accept_data checks for.
For global CPR we don't need to check the result, do it anyway for
consistency.